+++ rog-client/src/anime_dbus.rs
@@ -1,4 +1,5 @@
-use super::*;
+use crate::anime_matrix::AniMePacketType;
+use crate::{DBUS_IFACE, DBUS_NAME, DBUS_PATH};
 use dbus::channel::Sender;
 use dbus::{blocking::Connection, Message};
 use std::error::Error;